refactor(popup): extract inline editing logic into features/inline-edit
