+++ ui/popup/popup.js
@@ -5,9 +5,7 @@ import { fetchDoli } from './src/api/dolibarr.js';
 import { extractTextFromHtml, escapeHtml, formatLineBreaksForAttribute } from './src/utils/formatters.js';
 import { store } from './src/store/store.js';
 import { mapTicket } from './src/models/ticket.mapper.js';
-import { mapOpportunity } from './src/models/opportunity.mapper.js';
 import { renderTicketItemHtml } from './src/components/ticket.js';
-import { renderOppItemHtml } from './src/components/opportunity.js';
 
 class CustomSelect {
     constructor(selectElement) {
@@ -632,7 +630,144 @@ document.addEventListener('DOMContentLoaded', () => {
 
         // Fonction pour charger les dernières opportunités (Projets)
         
+        function renderOppItemHtml(project, doliBaseUrl, usersList, customOppDict, oppOriginDict, dolibarrNativeInputReasons) {
+            let subject = project.title || project.ref || "Projet sans titre";
+            if (subject.length > 50) subject = subject.substring(0, 50) + '...';
 
+            let statusColor = "#95a5a6";
+            const stat = String(project.statut || project.status || "0");
+            
+            let statusLabelText = stat;
+            const oppStatusMap = {
+                "0": "Brouillon",
+                "1": "Validé / Ouvert",
+                "2": "Clôturé"
+            };
+            if (oppStatusMap[stat]) statusLabelText = oppStatusMap[stat];
+            if (project.status_label) statusLabelText = project.status_label;
+            else if (project.statut_label) statusLabelText = project.statut_label;
+
+            if (stat === "0") statusColor = "#3498db"; // Brouillon
+            else if (stat === "1") statusColor = "#27ae60"; // Validé/Ouvert
+            else if (stat === "2") statusColor = "#7f8c8d"; // Clôturé
+
+            const projectRef = project.ref || `PROJ #${project.id}`;
+            
+            let dateCStr = "";
+            if (project.date_c) {
+                const d = new Date(project.date_c * 1000);
+                dateCStr = d.toLocaleDateString('fr-FR', { day: '2-digit', month: '2-digit', year: '2-digit' });
+            }
+            
+            let initials = "?";
+            if (project.user_author_id) {
+                const u = usersList.find(u => u.id == project.user_author_id);
+                if (u) {
+                    const parts = [u.firstname, u.lastname].filter(Boolean);
+                    if (parts.length >= 2) initials = parts[0].charAt(0).toUpperCase() + parts[1].charAt(0).toUpperCase();
+                    else if (parts.length === 1) initials = parts[0].substring(0, 2).toUpperCase();
+                    else if (u.login) initials = u.login.substring(0, 2).toUpperCase();
+                } else {
+                    initials = `U${project.user_author_id}`;
+                }
+            }
+
+            let amountDisplay = project.opp_amount ? new Intl.NumberFormat('fr-FR', { style: 'currency', currency: 'EUR', minimumFractionDigits: 0 }).format(project.opp_amount) : '';
+            let probDisplay = project.opp_percent ? `${Math.round(parseFloat(project.opp_percent))} %` : '';
+
+            const opts = project.array_options || {};
+            const oppNom = opts.options_reedcrm_lastname || '';
+            const oppPrenom = opts.options_reedcrm_firstname || '';
+            const oppTel = opts.options_projectphone || '';
+            const oppEmail = opts.options_reedcrm_email || '';
+            const oppWebsite = opts.options_reedcrm_website || opts.options_website || project.url || '';
+            const oppOriginRaw = opts.options_opporigin || opts.options_origine_opportunite || opts.options_origine || opts.options_origin || opts.options_source || opts.options_provenance || opts.options_prov || opts.options_opp_origin || opts.options_canal || '';
+            
+            let mappedOrigin = oppOriginRaw;
+            if (oppOriginRaw && customOppDict && customOppDict[oppOriginRaw]) {
+                mappedOrigin = customOppDict[oppOriginRaw];
+            } else if (oppOriginRaw && oppOriginDict && oppOriginDict[oppOriginRaw]) {
+                mappedOrigin = oppOriginDict[oppOriginRaw];
+            } else if (oppOriginRaw && dolibarrNativeInputReasons[oppOriginRaw]) {
+                mappedOrigin = dolibarrNativeInputReasons[oppOriginRaw];
+            }
+            
+            const oppOrigin = typeof mappedOrigin === 'string' ? mappedOrigin.charAt(0).toUpperCase() + mappedOrigin.slice(1).replace(/_/g, ' ') : mappedOrigin;
+
+            
+            const prenomVal = oppPrenom || "";
+            const nomVal = oppNom || "";
+            const telVal = oppTel || "";
+            const emailVal = oppEmail || "";
+            const websiteVal = oppWebsite || "";
+
+            let line1Html = `<div style="display: flex; align-items: center; gap: 4px;">` +
+                            `<span class="inline-editable ${!prenomVal ? 'placeholder-text' : ''}" data-field="options_reedcrm_firstname" data-pid="${project.id}" data-val="${prenomVal}" title="Cliquez pour modifier">${prenomVal || 'Prénom'}</span> ` +
+                            `<span class="inline-editable ${!nomVal ? 'placeholder-text' : ''}" data-field="options_reedcrm_lastname" data-pid="${project.id}" data-val="${nomVal}" title="Cliquez pour modifier">${nomVal || 'Nom'}</span>` +
+                            `</div>` +
+                            `<span class="rt-sep">&bull;</span>` +
+                            `<div style="display: flex; align-items: center; gap: 4px;">` +
+                            `<span class="inline-editable ${!telVal ? 'placeholder-text' : ''}" data-field="options_projectphone" data-pid="${project.id}" data-val="${telVal}" title="Cliquez pour modifier">${telVal || '0102030405'}</span>` +
+                            `<svg class="copy-icon" data-copy-target="tel" data-copy="${telVal}" width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" title="Copier le numéro"><rect x="9" y="9" width="13" height="13" rx="2" ry="2"></rect><path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1"></path></svg>` +
+                            `</div>`;
+
+            let displayEmail = emailVal.length > 45 ? emailVal.substring(0, 45) + '...' : emailVal;
+            let line2Html = `<div style="display: flex; align-items: center; gap: 4px;">` +
+                            `<span class="inline-editable ${!emailVal ? 'placeholder-text' : ''}" data-field="options_reedcrm_email" data-pid="${project.id}" data-val="${emailVal}" title="Cliquez pour modifier">${displayEmail || 'nomail@nomail.com'}</span>` +
+                            `<svg class="copy-icon" data-copy-target="email" data-copy="${emailVal}" width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" title="Copier l'email"><rect x="9" y="9" width="13" height="13" rx="2" ry="2"></rect><path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1"></path></svg>` +
+                            `</div>`;
+            
+            let line3Html = `<div class="rt-contact-line-web" style="display: flex; align-items: center; gap: 4px;">` +
+                            `<span class="inline-editable ${!websiteVal ? 'placeholder-text' : ''}" data-field="options_reedcrm_website" data-pid="${project.id}" data-val="${websiteVal}" title="Cliquez pour modifier">${websiteVal || 'https://www.website.com'}</span>` +
+                            (websiteVal ? ` <a href="${websiteVal.startsWith('http') ? websiteVal : 'https://' + websiteVal}" target="_blank" class="rt-contact-link"><svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"></path><polyline points="15 3 21 3 21 9"></polyline><line x1="10" y1="14" x2="21" y2="3"></line></svg></a>` : '') +
+                            `</div>`;
+            
+            let contactHtml = '';
+            if (line1Html !== '' || line2Html !== '' || line3Html !== '' || oppOrigin !== '') {
+                contactHtml = `<div class="rt-contact">`;
+                if (line1Html !== '') {
+                    contactHtml += `<div class="rt-contact-line1 rt-contact-line">${line1Html}</div>`;
+                }
+                if (line2Html !== '') {
+                    contactHtml += `<div class="rt-contact-line2 rt-contact-line" style="margin-top: 1px;">${line2Html}</div>`;
+                }
+                if (line3Html !== '') {
+                    contactHtml += `<div class="rt-contact-line-web rt-contact-line" style="margin-top: 1px;">${line3Html}</div>`;
+                }
+                if (oppOrigin !== '') {
+                    contactHtml += `<div class="rt-contact-line3" style="margin-top: 3px; display: flex; align-items: center; color: #475569; font-size: 11px;">
+                        <svg width="11" height="11" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" style="margin-right: 4px;"><path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path><polyline points="16 17 21 12 16 7"></polyline><line x1="21" y1="12" x2="9" y2="12"></line></svg>
+                        <span style="white-space: nowrap; overflow: hidden; text-overflow: ellipsis; flex: 1;" title="${oppOrigin}">${oppOrigin}</span>
+                    </div>`;
+                }
+                contactHtml += `</div>`;
+            }
+
+            const fullName = `${prenomVal} ${nomVal}`.trim();
+            const searchString = (projectRef + ' ' + subject + ' ' + fullName + ' ' + oppTel + ' ' + oppEmail).toLowerCase().replace(/['"]/g, '');
+
+            return `
+        <div id="opp-list-item-${project.id}" class="recent-ticket-item opp-list-item" style="align-items: flex-start;" data-search="${searchString}" data-date="${project.date_c || 0}" data-stat="${stat}">
+            <div class="rt-left">
+                <div class="rt-ref-group" style="display: flex; align-items: center; gap: 6px; white-space: nowrap;">
+                    <a href="${doliBaseUrl}/projet/card.php?id=${project.id}" target="_blank" class="rt-ref" title="Ouvrir le projet">${projectRef}</a>
+                    ${dateCStr ? `<span class="rt-sep">&bull;</span><div style="font-size: 10px; color: #888;">${dateCStr}</div>` : ''}
+                    ${initials !== "?" ? `<span class="rt-sep">&bull;</span><div style="font-size: 9px; background: #e2e8f0; color: #475569; padding: 1px 4px; border-radius: 4px;" title="Créé par">#${initials}</div>` : ''}
+                </div>
+                <div class="rt-subject" title="${subject.replace(/"/g, '')}" style="display: flex; align-items: center; gap: 6px; margin-top: 3px;">
+                    <div class="rt-status-dot" title="Statut: ${statusLabelText}" style="width: 8px; height: 8px; border-radius: 50%; background-color: ${statusColor}; flex-shrink: 0;"></div>
+                    <span style="white-space: nowrap; overflow: hidden; text-overflow: ellipsis; display: block; flex: 1; font-size: 13px; color: #334155; font-weight: 400;">${subject}</span>
+                </div>
+                ${contactHtml}
+            </div>
+            <div class="rt-right" style="display: flex; flex-direction: column; align-items: flex-end; justify-content: flex-start;">
+                <div class="rt-stats" style="font-size: 13px; font-weight: 400; display: flex; flex-direction: column; align-items: flex-end; gap: 2px; white-space: nowrap;">
+                    <span class="inline-editable ${!project.opp_percent ? 'placeholder-text' : ''}" data-field="opp_percent" data-pid="${project.id}" data-val="${project.opp_percent || ''}" title="Cliquez pour modifier le pourcentage" style="color: #1e293b;">${probDisplay || '0 %'}</span>
+                    <span class="inline-editable ${!project.opp_amount ? 'placeholder-text' : ''}" data-field="opp_amount" data-pid="${project.id}" data-val="${project.opp_amount || ''}" title="Cliquez pour modifier le montant" style="color: #1e293b;">${amountDisplay || '0 €'}</span>
+                </div>
+            </div>
+        </div>`;
+        }
 
         async function loadRecentOpportunities(apiUrl, token, limit = 10, entity, doliOppOnly = true, usersPromise = null, customDictMapStr = "") {
             recentOppContainer.classList.remove('hidden');
@@ -784,22 +919,18 @@ document.addEventListener('DOMContentLoaded', () => {
             }
         }
 
-        async function loadAllOpportunities(apiUrl, token, limit = 50, listCount, entity, oppOnly = true, isFullLoad = false) {
-            recentOppContainer.classList.remove('hidden');
-            const formOppContainer = document.getElementById('recent-opp-container-form');
-            if (formOppContainer) formOppContainer.classList.remove('hidden');
+        async function loadAllOpportunities(apiUrl, token, limit, listCount, entity, doliOppOnly, isBackground = false) {
+            const allOppList = document.getElementById('all-opp-list');
+            if (!allOppList) return;
             
-            const loaderHtml = `
-                <div class="loader-container">
-                    <div class="loader-spinner"></div>
-                    <div>${chrome.i18n.getMessage("popup_33") || "Chargement des opportunites..."}</div>
-                </div>
-            `;
-            oppList.innerHTML = loaderHtml;
-            const formOppList = document.getElementById('recent-opp-list-form');
-            if (formOppList) formOppList.innerHTML = loaderHtml;
-
-            const doliBaseUrl = apiUrl.replace(/\/api\/index\.php\/?$/, '');
+            if (!isBackground) {
+                allOppList.innerHTML = `
+                    <div class="loader-container">
+                        <div class="loader-spinner"></div>
+                        <div>${chrome.i18n.getMessage("popup_20")}</div>
+                    </div>
+                `;
+            }
 
             try {
                 const headers = {
@@ -810,63 +941,158 @@ document.addEventListener('DOMContentLoaded', () => {
                     headers['DOLAPIENTITY'] = String(entity).trim();
                 }
 
+                // Fetch locally with high limit for JS filtering
                 const response = await fetchDoli(`${apiUrl}/projects?sortfield=t.rowid&sortorder=DESC&limit=${limit}`, {
                     method: 'GET',
                     headers: headers
                 });
 
                 if (response.ok) {
-                    const textData = await response.text();
-                    const projects = textData.trim() ? JSON.parse(textData) : [];
-
+                    const projects = await response.json();
+                    
                     if (Array.isArray(projects) && projects.length > 0) {
-                        const sortedProjects = projects.sort((a, b) => b.date_c - a.date_c);
+                        let oppProjects = projects;
                         
-                        const mappedOpps = sortedProjects.map(p => mapOpportunity(p, store.state));
-                        store.setOpportunities(mappedOpps);
+                        if (doliOppOnly) {
+                            try {
+                                const activeProfile = getActiveProfile();
+                                const doliVersion = activeProfile && activeProfile.doliVersion ? activeProfile.doliVersion : "20.0";
+                                const isOldDoli = parseFloat(doliVersion) < 20;
 
-                        oppList.innerHTML = '';
-                        if (formOppList) formOppList.innerHTML = '';
-
-                        store.state.opportunities.forEach(mappedOpp => {
-                            const html = renderOppItemHtml(mappedOpp);
-                            
-                            const div = document.createElement('div');
-                            div.innerHTML = html.trim();
-                            const newCard = div.firstChild;
-                            oppList.appendChild(newCard);
-                            
-                            if (formOppList) {
-                                const formDiv = document.createElement('div');
-                                formDiv.innerHTML = html.trim();
-                                const newFormCard = formDiv.firstChild;
-                                formOppList.appendChild(newFormCard);
+                                oppProjects = projects.filter(p => {
+                                    if (isOldDoli && p.usage_bill_time === "1" && p.usage_opportunity !== "1") {
+                                        return false; 
+                                    }
+                                    return p.usage_opportunity == 1 || p.usage_opportunity === "1";
+                                });
+                            } catch(e) {
+                                oppProjects = projects.filter(p => p.usage_opportunity == 1 || p.usage_opportunity === "1");
                             }
-                        });
+                        }
 
-                        if (typeof initInlineEdit === 'function') {
-                            initInlineEdit(apiUrl, token, entity);
+                        if (oppProjects.length > 0) {
+                            if (!isBackground) allOppList.innerHTML = '';
+                            const sortedProjects = oppProjects.sort((a, b) => b.date_c - a.date_c);
+                            
+                            // Get users
+                            let usersList = [];
+                            try {
+                                const uRes = await fetchDoli(`${apiUrl}/users?limit=500`, { method: 'GET', headers: headers });
+                                if (uRes.ok) usersList = await uRes.json();
+                            } catch(e) {}
+                            
+                            // Dictionary maps (native)
+                            const dolibarrNativeInputReasons = {
+                                "1": "Campagne d'emailing",
+                                "2": "Campagne Fax",
+                                "3": "Campagne Publipostage",
+                                "4": "Campagne Téléphonique",
+                                "5": "Contact commercial",
+                                "6": "Contact entrant",
+                                "7": "Employé",
+                                "8": "Internet",
+                                "9": "Partenaire",
+                                "10": "Contact en boutique",
+                                "11": "Parrainage",
+                                "12": "Bouche à oreille"
+                            };
+
+                            let customOppDict = {};
+                            let oppOriginDict = {};
+                            
+                            // Get dictionaries
+                            try {
+                                const activeProfile = getActiveProfile();
+                                if (activeProfile && activeProfile.doliDictMap) {
+                                    activeProfile.doliDictMap.split(/\r?\n/).forEach(line => {
+                                        const parts = line.split(':');
+                                        if (parts.length >= 2) customOppDict[parts[0].trim()] = parts.slice(1).join(':').trim();
+                                    });
+                                }
+
+                                const efRes = await fetchDoli(`${apiUrl}/setup/extrafields`, { method: 'GET', headers: headers });
+                                if (efRes.ok) {
+                                    const efJson = await efRes.json();
+                                    let oField = null;
+                                    if (Array.isArray(efJson)) {
+                                        oField = efJson.find(f => f.name === 'opporigin' || f.name === 'origine_opportunite' || f.name === 'origine');
+                                    } else {
+                                        const pFields = efJson.project || efJson.projet || efJson;
+                                        oField = pFields.options_opporigin || pFields.opporigin || pFields.origine_opportunite || pFields.options_origine_opportunite;
+                                    }
+
+                                    if (oField && oField.param) {
+                                        if (typeof oField.param === 'string') {
+                                            oField.param.split(/\r?\n/).forEach(line => {
+                                                const parts = line.split(',');
+                                                if (parts.length >= 2) oppOriginDict[parts[0].trim()] = parts.slice(1).join(',').trim();
+                                            });
+                                        } else if (typeof oField.param === 'object') {
+                                            const paramsToMerge = oField.param.options || oField.param;
+                                            Object.keys(paramsToMerge).forEach(k => {
+                                                oppOriginDict[k.toString()] = paramsToMerge[k];
+                                            });
+                                        }
+                                    }
+                                }
+                            } catch(e) {}
+
+                            // Extract the domain
+                            let doliBaseUrl = apiUrl.replace('/api/index.php', '').replace('/api', '');
+
+                            let openCount = 0;
+                            let htmlToAppend = "";
+                            let renderedCount = document.querySelectorAll('.opp-list-item').length;
+                            sortedProjects.forEach((project, index) => {
+                                if (String(project.statut || project.status || "0") === "1") openCount++;
+                                
+                                // Prevent re-rendering if it already exists
+                                if (isBackground && document.getElementById(`opp-list-item-${project.id}`)) {
+                                    return;
+                                }
+
+                                const html = renderOppItemHtml(project, doliBaseUrl, usersList, customOppDict, oppOriginDict, dolibarrNativeInputReasons);
+                                
+                                // Build final HTML string directly to avoid DOMParser overhead
+                                let displayStyle = index >= listCount ? 'display: none;' : '';
+                                let visibilityClass = index >= listCount ? 'initially-hidden' : 'initially-visible';
+                                
+                                // We inject our classes and styles into the first div
+                                const modifiedHtml = html.replace('class="', `class="${visibilityClass} `).replace('style="', `style="${displayStyle} `);
+                                htmlToAppend += modifiedHtml;
+                                renderedCount++;
+                            });
+                            
+                            if (htmlToAppend) {
+                                allOppList.insertAdjacentHTML('beforeend', htmlToAppend);
+                            }
+                            
+                            if (typeof applyOppFilters === 'function') {
+                                applyOppFilters();
+                            }
+                            
+                            const countEl = document.getElementById('opp-count-total');
+                            if (countEl) {
+                                if (isBackground || limit >= 1000) {
+                                    countEl.textContent = openCount;
+                                } else {
+                                    countEl.innerHTML = `<div class="loader-spinner small" style="border-top-color: #c0392b; border-color: rgba(192,57,43,0.3); border-top-color: #c0392b; width: 10px; height: 10px; border-width: 2px;"></div><span style="font-size:10px;">chargement...</span>`;
+                                }
+                            }
+
+                        } else {
+                            allOppList.innerHTML = `<div style="text-align: center; color: #999;font-size: 11px; padding: 10px;">${chrome.i18n.getMessage('popup_js_110')}</div>`;
+                            const countEl = document.getElementById('opp-count-total');
+                            if (countEl) countEl.textContent = "0";
                         }
                     } else {
-                        const emptyHtml = `<div style="text-align: center; color: #999;font-size: 11px; padding: 10px;">Aucune opportunite recente.</div>`;
-                        oppList.innerHTML = emptyHtml;
-                        if (formOppList) formOppList.innerHTML = emptyHtml;
+                        allOppList.innerHTML = `<div style="text-align: center; color: #999;font-size: 11px; padding: 10px;">${chrome.i18n.getMessage('popup_js_110')}</div>`;
                     }
                 } else {
-                    oppList.innerHTML = '';
-                    if (formOppList) formOppList.innerHTML = '';
-                    const d = document.createElement('div'); d.style.cssText = "text-align: center; color: #e74c3c;font-size: 11px; padding: 10px;";
-                    d.textContent = `Erreur API (${response.status})`;
-                    oppList.appendChild(d);
-                    if (formOppList) formOppList.appendChild(d.cloneNode(true));
+                    allOppList.innerHTML = `<div style="text-align: center; color: #e74c3c;font-size: 11px; padding: 10px;">Erreur API (${response.status})</div>`;
                 }
             } catch (error) {
-                oppList.innerHTML = '';
-                if (formOppList) formOppList.innerHTML = '';
-                const d = document.createElement('div'); d.style.cssText = "text-align: center; color: #e74c3c;font-size: 11px; padding: 10px;";
-                d.textContent = `Erreur JS: ${error.message}`;
-                oppList.appendChild(d);
-                if (formOppList) formOppList.appendChild(d.cloneNode(true));
+                allOppList.innerHTML = `<div style="text-align: center; color: #e74c3c;font-size: 11px; padding: 10px;">Erreur JS: ${error.message}</div>`;
             }
         }
 
@@ -2328,495 +2554,4 @@ document.addEventListener('DOMContentLoaded', () => {
     }
 });
 
-document.addEventListener('dblclick', (e) => {
-    if (e.target.classList.contains('copy-able')) {
-        const textToCopy = e.target.getAttribute('data-copy');
-        if (textToCopy) {
-            navigator.clipboard.writeText(textToCopy).then(() => {
-                const originalText = e.target.textContent;
-                e.target.textContent = chrome.i18n.getMessage('popup_js_133');
-                e.target.style.color = '#27ae60';
-                e.target.style.fontWeight = 'bold';
-                setTimeout(() => {
-                    e.target.textContent = originalText;
-                    e.target.style.color = '';
-                    e.target.style.fontWeight = '';
-                }, 1000);
-            });
-            window.getSelection().removeAllRanges();
-        }
-    }
-});
-
-
-document.addEventListener('click', async (e) => {
-    const copyIcon = e.target.closest('.copy-icon');
-    if (copyIcon) {
-        e.preventDefault();
-        e.stopPropagation();
-        const text = copyIcon.getAttribute('data-copy');
-        if (text) {
-            try {
-                await navigator.clipboard.writeText(text);
-                const origColor = copyIcon.style.color;
-                copyIcon.style.color = '#10b981'; // Green
-                setTimeout(() => { copyIcon.style.color = origColor; }, 1000);
-            } catch (err) {}
-        }
-        return;
-    }
-
-    const editable = e.target.closest('.inline-editable');
-    if (editable) {
-        if (editable.querySelector('input') || editable.querySelector('select') || editable.querySelector('textarea')) return;
-        
-        e.preventDefault();
-        e.stopPropagation();
-        
-        const currentValue = editable.getAttribute('data-val') || '';
-        const projectId = editable.getAttribute('data-pid');
-        const fieldName = editable.getAttribute('data-field');
-        
-        let input;
-        if (fieldName === 'fk_statut' || fieldName === 'severity_code' || fieldName === 'fk_user_assign') {
-            input = document.createElement('select');
-            input.id = 'inline_edit_' + fieldName + '_' + projectId;
-            input.className = 'inline-edit-input inline-edit-select';
-            
-            if (fieldName === 'fk_statut') {
-                const statuses = {
-                    "0": "Non lu",
-                    "1": "Lu",
-                    "2": "Assigné",
-                    "3": "En cours",
-                    "4": "En attente de retour",
-                    "5": "En attente",
-                    "8": "Fermé (Résolu)",
-                    "9": "Annulé"
-                };
-                for (const [val, label] of Object.entries(statuses)) {
-                    const opt = document.createElement('option');
-                    opt.value = val;
-                    opt.textContent = label.length > 15 ? label.substring(0, 15) + '..' : label;
-                    if (String(val) === String(currentValue)) opt.selected = true;
-                    input.appendChild(opt);
-                }
-            } else if (fieldName === 'severity_code') {
-                const severities = {
-                    "LOW": "Basse",
-                    "NORMAL": "Normale",
-                    "HIGH": "Haute",
-                    "BLOCKING": "Bloquante"
-                };
-                for (const [val, label] of Object.entries(severities)) {
-                    const opt = document.createElement('option');
-                    opt.value = val;
-                    opt.textContent = label.length > 15 ? label.substring(0, 15) + '..' : label;
-                    if (String(val) === String(currentValue)) opt.selected = true;
-                    input.appendChild(opt);
-                }
-            } else if (fieldName === 'fk_user_assign') {
-                const optEmpty = document.createElement('option');
-                optEmpty.value = "";
-                optEmpty.textContent = "Non assigné";
-                input.appendChild(optEmpty);
-                if (window.usersList && window.usersList.length > 0) {
-                    window.usersList.forEach(u => {
-                        const opt = document.createElement('option');
-                        opt.value = u.id;
-                        let userLabel = (u.firstname || '') + ' ' + (u.lastname || u.login || '');
-                        opt.textContent = userLabel.length > 15 ? userLabel.substring(0, 15) + '..' : userLabel;
-                        if (String(u.id) === String(currentValue)) opt.selected = true;
-                        input.appendChild(opt);
-                    });
-                }
-            }
-        } else if (fieldName === 'message') {
-            input = document.createElement('textarea');
-            input.id = 'inline_edit_' + fieldName + '_' + projectId;
-            input.name = 'inline_edit_' + fieldName;
-            input.className = 'inline-edit-input';
-            input.value = currentValue;
-            input.style.height = '80px';
-            input.style.resize = 'vertical';
-        } else {
-            input = document.createElement('input');
-            input.type = 'text';
-            input.id = 'inline_edit_' + fieldName + '_' + projectId;
-            input.name = 'inline_edit_' + fieldName;
-            input.className = 'inline-edit-input';
-            input.value = currentValue;
-        }
-        
-        const originalNodes = Array.from(editable.childNodes).map(n => n.cloneNode(true));
-        const restoreOriginal = (el) => {
-            el.textContent = '';
-            originalNodes.forEach(n => el.appendChild(n.cloneNode(true)));
-        };
-        const originalClass = editable.className;
-        
-        editable.classList.add('is-editing');
-        editable.innerHTML = '';
-        editable.appendChild(input);
-        input.focus();
-        if (input.tagName === 'INPUT') {
-            input.select();
-        }
-        
-        let isSaving = false;
-        
-        const saveEdit = async () => {
-            if (isSaving) return;
-            isSaving = true;
-            
-            let newValue = input.value.trim();
-            
-            const showErrorInline = (msg) => {
-                editable.style.transition = 'color 0.3s ease';
-                editable.style.color = '#ef4444'; // Rouge vif
-                editable.innerHTML = msg;
-                setTimeout(() => {
-                    editable.style.color = '';
-                    editable.style.transition = '';
-                    restoreOriginal(editable);
-                    editable.className = originalClass;
-                    editable.classList.remove('is-editing');
-                    // On ne remet pas isSaving à false car on a restauré l'état initial (plus d'input)
-                }, 3000);
-            };
-
-            if (fieldName === 'options_projectphone' && newValue !== '') {
-                if (/[a-zA-Z]/.test(newValue)) {
-                    showErrorInline(chrome.i18n.getMessage('popup_title_42') || "Veuillez saisir un numéro valide");
-                    return;
-                }
-                
-                // Suppression de tout ce qui n'est pas chiffre ou '+'
-                newValue = newValue.replace(/[^\d+]/g, '');
-                
-                if (newValue.length > 0 && (newValue.length < 9 || newValue.length > 15)) {
-                    showErrorInline(chrome.i18n.getMessage('popup_title_42') || "Veuillez saisir un numéro valide");
-                    return;
-                }
-                
-                // Formatage français basique (0X XX XX XX XX) si 10 chiffres commençant par 0
-                if (/^0[1-9]\d{8}$/.test(newValue)) {
-                    newValue = newValue.replace(/(\d{2})(?=\d)/g, '$1 ');
-                }
-            }
-            
-            if (fieldName === 'options_reedcrm_email' && newValue !== '') {
-                const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-                if (!emailRegex.test(newValue)) {
-                    showErrorInline(chrome.i18n.getMessage('popup_js_err_email') || "Email invalide");
-                    return;
-                }
-            }
-            
-            if (fieldName === 'options_reedcrm_website' && newValue !== '') {
-                const websiteRegex = /^(https?:\/\/)?([\da-z\.-]+)\.([a-z\.]{2,6})([\/\w \.-]*)*\/?$/i;
-                if (!websiteRegex.test(newValue)) {
-                    showErrorInline(chrome.i18n.getMessage('popup_title_43') || "Exemple de domaine valide: monsite.com");
-                    return;
-                }
-                if (!newValue.startsWith('http://') && !newValue.startsWith('https://')) {
-                    newValue = 'https://' + newValue;
-                }
-            }
-            
-            if ((fieldName === 'opp_percent' || fieldName === 'progress') && newValue !== '') {
-                if (/[^\d.,]/.test(newValue)) {
-                    showErrorInline(chrome.i18n.getMessage('popup_js_err_percent') || "Le pourcentage doit être entre 0 et 100");
-                    return;
-                }
-                newValue = newValue.replace(',', '.');
-                const percentVal = parseFloat(newValue);
-                if (isNaN(percentVal) || percentVal < 0 || percentVal > 100) {
-                    showErrorInline(chrome.i18n.getMessage('popup_js_err_percent') || "Le pourcentage doit être entre 0 et 100");
-                    return;
-                }
-                newValue = percentVal.toString();
-            }
-            
-            if (fieldName === 'opp_amount' && newValue !== '') {
-                if (/[^\d.,\s]/.test(newValue)) {
-                    showErrorInline(chrome.i18n.getMessage('popup_js_err_amount') || "Montant invalide (chiffres uniquement)");
-                    return;
-                }
-                newValue = newValue.replace(/\s/g, '').replace(',', '.');
-                const amountVal = parseFloat(newValue);
-                if (isNaN(amountVal) || amountVal < 0) {
-                    showErrorInline(chrome.i18n.getMessage('popup_js_err_amount') || "Montant invalide (chiffres uniquement)");
-                    return;
-                }
-                newValue = amountVal.toString();
-            }
-            
-            if (newValue === currentValue) {
-                restoreOriginal(editable);
-                editable.className = originalClass;
-                editable.classList.remove('is-editing');
-                return;
-            }
-            
-            input.disabled = true;
-            input.style.opacity = '0.5';
-            
-            try {
-                const profiles = await new Promise(resolve => chrome.storage.sync.get('doliProfiles', data => resolve(data.doliProfiles || [])));
-                let activeProfileId = null;
-                await new Promise(resolve => chrome.storage.sync.get('doliActiveProfileId', data => { activeProfileId = data.doliActiveProfileId; resolve(); }));
-                
-                const profile = activeProfileId ? profiles.find(p => p.id === activeProfileId) : profiles[0];
-                if (!profile) throw new Error("No profile found");
-                
-                const apiUrl = profile.doliUrl;
-                const token = profile.doliApiToken;
-                
-                let payload = {};
-                if (fieldName.startsWith('options_')) {
-                    payload = {
-                        array_options: {
-                            [fieldName]: newValue
-                        }
-                    };
-                } else {
-                    payload = {
-                        [fieldName]: newValue
-                    };
-                    if (fieldName === 'fk_statut') {
-                        payload.statut = parseInt(newValue, 10);
-                        payload.status = parseInt(newValue, 10);
-                        payload.fk_statut = parseInt(newValue, 10);
-                        if (String(newValue) === '8') {
-                            payload.progression = 100;
-                            payload.progress = 100;
-                        }
-                    } else if (fieldName === 'progress') {
-                        payload.progression = parseInt(newValue, 10);
-                    }
-                }
-                
-                let endpointUrl = `${apiUrl}/projects/${projectId}`;
-                if (['fk_statut', 'severity_code', 'progress', 'fk_user_assign', 'subject', 'message'].includes(fieldName)) {
-                    endpointUrl = `${apiUrl}/tickets/${projectId}`;
-                }
-                
-                const res = await fetchDoli(endpointUrl, {
-                    method: 'PUT',
-                    headers: {
-                        'DOLAPIKEY': token,
-                        'Content-Type': 'application/json'
-                    },
-                    body: JSON.stringify(payload)
-                });
-                
-                if (res.ok) {
-                    editable.setAttribute('data-val', newValue);
-                    let displayValue = newValue;
-                    if (input.tagName === 'SELECT') {
-                        const selOpt = input.options[input.selectedIndex];
-                        displayValue = selOpt ? selOpt.text : newValue;
-                    }
-
-                    if (fieldName === 'fk_statut') {
-                        restoreOriginal(editable);
-                        const label = editable.querySelector('.tc-status-label');
-                        if (label) label.textContent = displayValue;
-                        const dot = editable.querySelector('.tc-status-dot');
-                        if (dot) {
-                            let statusColor = "#95a5a6";
-                            const stat = String(newValue);
-                            if (stat === "0") statusColor = "#e74c3c";
-                            else if (stat === "1") statusColor = "#3498db";
-                            else if (["2", "3", "4", "5", "6", "7"].includes(stat)) statusColor = "#f39c12";
-                            else if (stat === "8") statusColor = "#27ae60";
-                            else if (stat === "9") statusColor = "#7f8c8d";
-                            dot.style.backgroundColor = statusColor;
-                        }
-                        
-                        // Met à jour la progression à 100% si on clôture le ticket
-                        if (String(newValue) === '8') {
-                            const ticketCard = editable.closest('.recent-ticket-item');
-                            if (ticketCard) {
-                                const progressEl = ticketCard.querySelector('[data-field="progress"]');
-                                if (progressEl) {
-                                    progressEl.setAttribute('data-val', '100');
-                                    progressEl.textContent = '100%';
-                                }
-                            }
-                        }
-                    } else if (fieldName === 'fk_user_assign') {
-                        if (!newValue) {
-                            editable.textContent = '?';
-                        } else {
-                            const matchedUser = window.usersList ? window.usersList.find(u => String(u.id) === String(newValue)) : null;
-                            const parts = displayValue.split(' ');
-                            let initials = parts.length > 1 ? parts[0].charAt(0).toUpperCase() + parts[1].charAt(0).toUpperCase() : displayValue.substring(0, 2).toUpperCase();
-                            if (matchedUser && matchedUser.photo && matchedUser.photo.trim() !== '') {
-                                const img = document.createElement('img');
-                                let photoUrl = matchedUser.photo.trim();
-                                if (!photoUrl.startsWith('http') && !photoUrl.startsWith('//')) {
-                                    photoUrl = `${apiUrl.replace('/api/index.php', '')}/document.php?modulepart=user&file=${encodeURIComponent(photoUrl)}`;
-                                }
-                                img.src = photoUrl;
-                                img.alt = initials;
-                                img.onerror = () => {
-                                    editable.textContent = initials;
-                                };
-                                editable.textContent = '';
-                                editable.appendChild(img);
-                            } else {
-                                editable.textContent = initials;
-                            }
-                        }
-                    } else if (fieldName === 'severity_code') {
-                        editable.textContent = displayValue;
-                    } else if (fieldName === 'progress') {
-                        editable.textContent = newValue + '%';
-                    } else {
-                        if (!newValue) {
-                            editable.classList.add('placeholder-text');
-                            if (fieldName === 'options_reedcrm_firstname') displayValue = 'Prénom';
-                            else if (fieldName === 'options_reedcrm_lastname') displayValue = 'Nom';
-                            else if (fieldName === 'options_projectphone') displayValue = '0102030405';
-                            else if (fieldName === 'options_reedcrm_email') displayValue = 'nomail@nomail.com';
-                            else if (fieldName === 'options_reedcrm_website') displayValue = 'https://www.website.com';
-                            else if (fieldName === 'opp_percent') displayValue = '0 %';
-                            else if (fieldName === 'opp_amount') displayValue = '0 €';
-                        } else {
-                            editable.classList.remove('placeholder-text');
-                            if (fieldName === 'opp_percent') {
-                                displayValue = `${Math.round(parseFloat(newValue))} %`;
-                            } else if (fieldName === 'opp_amount') {
-                                displayValue = new Intl.NumberFormat('fr-FR', { style: 'currency', currency: 'EUR', minimumFractionDigits: 0 }).format(parseFloat(newValue));
-                            } else if (fieldName === 'options_reedcrm_website') {
-                                displayValue = newValue;
-                                if (editable.tagName === 'A') {
-                                    let safeHref = '#';
-                                    try {
-                                        const parsedUrl = new URL(newValue, window.location.origin);
-                                        if (parsedUrl.protocol === 'http:' || parsedUrl.protocol === 'https:') {
-                                            safeHref = parsedUrl.href;
-                                        }
-                                    } catch (e) {}
-                                    editable.setAttribute('href', safeHref);
-                                }
-                            }
-                        }
-                        
-                        let iconNode = null;
-                        for (const n of originalNodes) {
-                            if (n.nodeName === 'I') { iconNode = n.cloneNode(true); break; }
-                            if (n.querySelector && n.querySelector('i')) { iconNode = n.querySelector('i').cloneNode(true); break; }
-                        }
-                        
-                        editable.textContent = '';
-                        if (iconNode) {
-                            editable.appendChild(iconNode);
-                            editable.appendChild(document.createTextNode(' ' + displayValue));
-                        } else {
-                            editable.textContent = displayValue;
-                        }
-                    }
-                    
-                    const contactLine = editable.closest('.rt-contact-line');
-                    if (contactLine) {
-                        let targetBtn = null;
-                        if (fieldName === 'options_projectphone') targetBtn = contactLine.querySelector('[data-copy-target="tel"]');
-                        if (fieldName === 'options_reedcrm_email') targetBtn = contactLine.querySelector('[data-copy-target="email"]');
-                        if (targetBtn) targetBtn.setAttribute('data-copy', newValue);
-                        
-                        if (fieldName === 'options_reedcrm_website') {
-                            const linkEl = contactLine.querySelector('.rt-contact-link');
-                            if (newValue) {
-                                const href = newValue.startsWith('http') ? newValue : 'https://' + newValue;
-                                if (linkEl) {
-                                    linkEl.href = href;
-                                } else {
-                                    const newLink = document.createElement('a');
-                                    newLink.href = href;
-                                    newLink.target = "_blank";
-                                    newLink.className = "rt-contact-link";
-                                    newLink.style.marginLeft = "2px";
-                                    newLink.innerHTML = `<svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"></path><polyline points="15 3 21 3 21 9"></polyline><line x1="10" y1="14" x2="21" y2="3"></line></svg>`;
-                                    editable.parentNode.insertBefore(newLink, editable.nextSibling);
-                                }
-                            } else {
-                                if (linkEl) linkEl.remove();
-                            }
-                        }
-                    }
-                    
-                    // Animation : passage au vert puis retour à la normale
-                    editable.style.transition = 'color 0.5s ease-out';
-                    editable.style.color = '#10b981';
-                    setTimeout(() => {
-                        editable.style.color = '';
-                        setTimeout(() => {
-                            editable.style.transition = '';
-                        }, 500);
-                    }, 1000);
-                } else {
-                    let errStr = "API error";
-                    try {
-                        const errJson = await res.json();
-                        errStr = errJson.error ? (errJson.error.message || JSON.stringify(errJson.error)) : JSON.stringify(errJson);
-                    } catch(e) {
-                        errStr = e.message || "API error";
-                    }
-                    throw new Error(errStr);
-                }
-            } catch (err) {
-                console.error(err);
-                alert(chrome.i18n.getMessage('popup_js_err_save') + err.message);
-                restoreOriginal(editable);
-                editable.className = originalClass;
-                editable.classList.remove('is-editing');
-            } finally {
-                editable.classList.remove('is-editing');
-            }
-        };
-        
-        if (input.tagName === 'SELECT') {
-            input.addEventListener('change', saveEdit);
-            const outsideClickListener = (evt) => {
-                if (!editable.contains(evt.target)) {
-                    document.removeEventListener('click', outsideClickListener);
-                    if (editable.classList.contains('is-editing') && !isSaving) {
-                        restoreOriginal(editable);
-                        editable.className = originalClass;
-                        editable.classList.remove('is-editing');
-                    }
-                }
-            };
-            setTimeout(() => document.addEventListener('click', outsideClickListener), 100);
-        } else {
-            input.addEventListener('blur', saveEdit);
-        }
-        input.addEventListener('keydown', (evt) => {
-            if (evt.key === 'Enter') {
-                if (input.tagName === 'TEXTAREA') {
-                    if (evt.ctrlKey || evt.metaKey) saveEdit();
-                    // Sinon on laisse faire le saut de ligne natif
-                } else {
-                    saveEdit();
-                }
-            } else if (evt.key === 'Tab') {
-                evt.preventDefault();
-                saveEdit().then(() => {
-                    const allEditables = Array.from(document.querySelectorAll('.inline-editable'));
-                    const currentIndex = allEditables.indexOf(editable);
-                    if (currentIndex > -1 && currentIndex < allEditables.length - 1) {
-                        allEditables[currentIndex + 1].click();
-                    }
-                });
-            } else if (evt.key === 'Escape') {
-                isSaving = true;
-                restoreOriginal(editable);
-                editable.className = originalClass;
-                editable.classList.remove('is-editing');
-            }
-        });
-    }
-});
+    initInlineEdit();
